+++ tests/test-helpers/app7-port/option.R
@@ -1,0 +1,13 @@
+library(shiny)
+
+options(shiny.port = 7777)
+
+ui <- fluidPage(
+
+)
+
+server <- function(input, output, session) {
+
+}
+
+shinyApp(ui, server)