+++ NEWS.md
@@ -10,6 +10,8 @@ shiny 1.5.0.9000
 
 * Fixed #2859: `renderPlot()` wasn't correctly setting `showtext::showtext_opts()`'s `dpi` setting with the correct resolution on high resolution displays; which means, if the font was rendered by showtext, font sizes would look smaller than they should on such displays. (#2941)
 
+* Fixed #1942: Calling `runApp("app.R")` ignored options passed into `shinyApp()`.
+
 
 shiny 1.5.0
 ===========